requirements.txt
python-dotenv==1.2.2
python-docx==1.2.0
openai==2.26.0
